Move spec ingest IO to shared example
